ZOOKEEPER-2261. Refactor ZooKeeperServerTest class and move existing tests to proper places
